requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==3.0.1
WTForms==3.1.1
SQLAlchemy==2.0.30
bcrypt==4.1.2
stripe==7.8.0
python-dotenv==1.0.0
email-validator==2.1.0
